feat(ui, config): implement data-driven dynamic camera test modules
- Migrate `CameraConfigPage` to a declarative `.ui` form to simplify layout and validation logic.
- Add `associatedFocalId` and `associatedLightId` optional hardware mappings to `camera_config_t`, eliminating hardcoded topologies.
- Move `CameraStreamWidget` and `DoubleRatioSlider` to the shared `core/` library for cross-application reuse.
- Refactor `CamerasTestView` to act as a dynamic grid, instantiating generic `CameraTestModule` UIs strictly from configuration data.
- Update `IFocal` and `ILightingSystem` to use percentage-based `setValueFraction(double)` instead of explicit 12-bit integer limits to decouple the UI from hardware specifics.

+++ apps/tester/include/Services/FocalTestService.h
@@ -33,7 +33,7 @@ namespace Kub3::Tools::Tester
 
         // --- Focal Specific API ---
         void toggleFocal(const std::string &focalId, bool enabled);
-        void updateFocalValue(const std::string &focalId, uint16_t value);
+        void updateFocalValue(const std::string &focalId, double value);
 
         [[nodiscard]] const std::vector<std::string> &getAvailableFocals() const;
